refactor(payment): use admin-managed FX rates
