+++ backend/cmd/server/wire_gen_test.go
@@ -75,7 +75,6 @@ func TestProvideCleanup_WithMinimalDependencies_NoPanic(t *testing.T) {
 		nil, // openAIGateway
 		nil, // scheduledTestRunner
 		nil, // backupSvc
-		nil, // paymentFXSync
 		nil, // paymentOrderExpiry
 		nil, // channelMonitorRunner
 	)
