+++ backend/internal/handler/admin/setting_handler.go
@@ -240,9 +240,6 @@ func (h *SettingHandler) GetSettings(c *gin.Context) {
 		PaymentAllowedCurrencies:               paymentCfg.AllowedPaymentCurrencies,
 		PaymentManualFXRates:                   paymentCfg.ManualFXRates,
 		PaymentFXStatus:                        paymentFXStatusDTO(paymentCfg.FXStatus),
-		PaymentFXAutoSyncEnabled:               paymentCfg.FXAutoSyncEnabled,
-		PaymentFXAutoSyncProvider:              paymentCfg.FXAutoSyncProvider,
-		PaymentFXAutoSyncIntervalSeconds:       paymentCfg.FXAutoSyncIntervalSec,
 		PaymentCancelRateLimitEnabled:          paymentCfg.CancelRateLimitEnabled,
 		PaymentCancelRateLimitMax:              paymentCfg.CancelRateLimitMax,
 		PaymentCancelRateLimitWindow:           paymentCfg.CancelRateLimitWindow,
@@ -489,9 +486,6 @@ type UpdateSettingsRequest struct {
 	PaymentLedgerCurrency            *string  `json:"payment_ledger_currency"`
 	PaymentAllowedCurrencies         []string `json:"payment_allowed_currencies"`
 	PaymentManualFXRates             *string  `json:"payment_manual_fx_rates"`
-	PaymentFXAutoSyncEnabled         *bool    `json:"payment_fx_auto_sync_enabled"`
-	PaymentFXAutoSyncProvider        *string  `json:"payment_fx_auto_sync_provider"`
-	PaymentFXAutoSyncIntervalSeconds *int     `json:"payment_fx_auto_sync_interval_seconds"`
 	PaymentFXRatesStaleAfterSeconds  *int     `json:"payment_fx_rates_stale_after_seconds"`
 
 	// Cancel rate limit
@@ -1445,9 +1439,6 @@ func (h *SettingHandler) UpdateSettings(c *gin.Context) {
 			LedgerCurrency:            req.PaymentLedgerCurrency,
 			AllowedPaymentCurrencies:  req.PaymentAllowedCurrencies,
 			ManualFXRates:             req.PaymentManualFXRates,
-			FXAutoSyncEnabled:         req.PaymentFXAutoSyncEnabled,
-			FXAutoSyncProvider:        req.PaymentFXAutoSyncProvider,
-			FXAutoSyncIntervalSec:     req.PaymentFXAutoSyncIntervalSeconds,
 			FXRatesStaleAfterSeconds:  req.PaymentFXRatesStaleAfterSeconds,
 			CancelRateLimitEnabled:    req.PaymentCancelRateLimitEnabled,
 			CancelRateLimitMax:        req.PaymentCancelRateLimitMax,
@@ -1627,9 +1618,6 @@ func (h *SettingHandler) UpdateSettings(c *gin.Context) {
 		PaymentAllowedCurrencies:               updatedPaymentCfg.AllowedPaymentCurrencies,
 		PaymentManualFXRates:                   updatedPaymentCfg.ManualFXRates,
 		PaymentFXStatus:                        paymentFXStatusDTO(updatedPaymentCfg.FXStatus),
-		PaymentFXAutoSyncEnabled:               updatedPaymentCfg.FXAutoSyncEnabled,
-		PaymentFXAutoSyncProvider:              updatedPaymentCfg.FXAutoSyncProvider,
-		PaymentFXAutoSyncIntervalSeconds:       updatedPaymentCfg.FXAutoSyncIntervalSec,
 		PaymentCancelRateLimitEnabled:          updatedPaymentCfg.CancelRateLimitEnabled,
 		PaymentCancelRateLimitMax:              updatedPaymentCfg.CancelRateLimitMax,
 		PaymentCancelRateLimitWindow:           updatedPaymentCfg.CancelRateLimitWindow,
@@ -1676,8 +1664,7 @@ func hasPaymentFields(req UpdateSettingsRequest) bool {
 		req.PaymentProductNameSuffix != nil || req.PaymentHelpImageURL != nil ||
 		req.PaymentHelpText != nil || req.PaymentLedgerCurrency != nil ||
 		req.PaymentAllowedCurrencies != nil || req.PaymentManualFXRates != nil ||
-		req.PaymentFXAutoSyncEnabled != nil || req.PaymentFXAutoSyncProvider != nil ||
-		req.PaymentFXAutoSyncIntervalSeconds != nil || req.PaymentFXRatesStaleAfterSeconds != nil ||
+		req.PaymentFXRatesStaleAfterSeconds != nil ||
 		req.PaymentCancelRateLimitEnabled != nil ||
 		req.PaymentCancelRateLimitMax != nil || req.PaymentCancelRateLimitWindow != nil ||
 		req.PaymentCancelRateLimitUnit != nil || req.PaymentCancelRateLimitMode != nil
