+++ deploy/.env.example
@@ -103,18 +103,11 @@ RUN_MODE=standard
 TZ=Asia/Shanghai
 
 # -----------------------------------------------------------------------------
-# Payment FX Provider (Optional, required for production auto-sync)
+# Payment FX rates
 # -----------------------------------------------------------------------------
-# Real FX provider selected for Phase 5: Open Exchange Rates.
-# Keep internal ledger USD; provider returns local units per USD and the app stores
-# payment-currency -> USD ledger rates in PAYMENT_MANUAL_FX_RATES_JSON.
-# Enable auto-sync from Admin Payment Settings with provider=openexchangerates.
-PAYMENT_FX_OPEN_EXCHANGE_RATES_APP_ID=
-PAYMENT_FX_OPEN_EXCHANGE_RATES_BASE_URL=https://openexchangerates.org/api/latest.json
-PAYMENT_FX_OPEN_EXCHANGE_RATES_TIMEOUT_SECONDS=10
-PAYMENT_FX_SYNC_MAX_ATTEMPTS=3
-PAYMENT_FX_SYNC_INITIAL_BACKOFF_SECONDS=2
-PAYMENT_FX_SYNC_MAX_BACKOFF_SECONDS=30
+# Payment FX rates are configured by Admin and stored in DB settings.
+# Do not configure an external FX API here. Checkout uses the admin-saved
+# payment-currency -> USD ledger rates snapshot at quote/order creation.
 
 # -----------------------------------------------------------------------------
 # PostgreSQL Configuration (REQUIRED)
