+++ frontend/src/api/admin/settings.ts
@@ -463,9 +463,6 @@ export interface SystemSettings {
   payment_allowed_currencies: string[];
   payment_manual_fx_rates: Record<string, number>;
   payment_fx_status: PaymentFXStatus;
-  payment_fx_auto_sync_enabled: boolean;
-  payment_fx_auto_sync_provider: string;
-  payment_fx_auto_sync_interval_seconds: number;
   payment_cancel_rate_limit_enabled: boolean;
   payment_cancel_rate_limit_max: number;
   payment_cancel_rate_limit_window: number;
@@ -638,9 +635,6 @@ export interface UpdateSettingsRequest {
   payment_ledger_currency?: string;
   payment_allowed_currencies?: string[];
   payment_manual_fx_rates?: string;
-  payment_fx_auto_sync_enabled?: boolean;
-  payment_fx_auto_sync_provider?: string;
-  payment_fx_auto_sync_interval_seconds?: number;
   payment_fx_rates_stale_after_seconds?: number;
   payment_cancel_rate_limit_enabled?: boolean;
   payment_cancel_rate_limit_max?: number;
